gadget_areafile {mfdb}	R Documentation

Gadget area files

Description

Structures representing a GADGET area file

Usage

gadget_areafile(size, temperature, area = attr(size, 'area'))

Arguments

size
data.frame as produced by mfdb_area_size

temperature
data.frame as produced by mfdb_area_temperature

area
Optional. mfdb_group that you used to specify area. By default pulls it from annotations on the size object.

Details

Once formed, you can then use gadget_dir_write to write this out to a GADGET areafile.

Examples

# Any example could be added to a file with the following:-
## Not run: 
gd <- gadget_directory('./out')
agg_data <- gadget_areafile(
    mfdb_area_size(...),
    mfdb_area_temperature(...))
gadget_dir_write(gd, agg_data[[1]])

## End(Not run)

[Package mfdb version 4.0-0 Index]