MareFrame DB unaggregated data

Tell mfdb functions not to aggregate this column, just return all values.

mfdb_unaggregated(omitNA = FALSE, like = c(), not_like = c())

Arguments

omitNA
Skip over rows where column is NA

like
Vector of SQL like expressions to check column against

not_like
Vector of SQL like expressions to check column against

Details

SQL like expessions can use the wildcards "_" to match any character and "

Value

An mfdb_aggregate object that can then be used in querying functions such as mfdb_sample_count

Examples

# All vessels with a name ending with 'e' or 'd'
mfdb_unaggregated(like = c("%e", "%d"))

#> $omitNA
#> [1] FALSE
#> 
#> $like
#> [1] "%e" "%d"
#> 
#> $not_like
#> NULL
#> 
#> attr(,"class")
#> [1] "mfdb_unaggregated" "mfdb_aggregate"