MareFrame DB Dump / Restore

Dump / Restore entire case studies.

mfdb_cs_dump(mdb, out_location)
mfdb_cs_restore(mdb, in_location)

Arguments

mdb

(required) A database connection created by mfdb()

in_location, out_location

(required) A filesystem directory or '.tar.gz' file to dump / restore database contents.

Details

Deprecated: These commands aren't strictly necessary any more. In most situations it will be easier to use Postgres' pg_dump and pg_restore. See mfdb-package for some examples of how to do it. These functions don't offer much more functionality and much slower.

mfdb_cs_dump copies all data from the database/case-study that mdb is connected to, and writes it out to files in out_location. If this ends with 'tar.gz', then all files will be put into a tarball with the name

mfdb_cs_restore will remove any case-study data, and replaces it with the content of in_location, a directory or tarball.

Value

NULL

Examples

# Copy data from one database to another, note they don't have to be the same type
mdb_out <- mfdb(tempfile(fileext = '.sqlite'))
#> 2026-01-22 15:11:08.68188 INFO:mfdb:Creating schema from scratch
#> 2026-01-22 15:11:08.770192 INFO:mfdb:Taxonomy market_category no updates to make
#> 2026-01-22 15:11:08.876586 INFO:mfdb:Schema up-to-date
mdb_in <- mfdb(tempfile(fileext = '.duckdb'))
#> 2026-01-22 15:11:08.935562 INFO:mfdb:Creating schema from scratch
#> 2026-01-22 15:11:09.361754 INFO:mfdb:Taxonomy market_category no updates to make
#> 2026-01-22 15:11:09.927677 INFO:mfdb:Schema up-to-date

dump_path <- tempfile(fileext='.tar.gz')
mfdb_cs_dump(mdb_out, dump_path)
#> 2026-01-22 15:11:09.92982 INFO:mfdb:Dumping table institute
#> 2026-01-22 15:11:09.931411 INFO:mfdb:Dumping table bait_type
#> 2026-01-22 15:11:09.932661 INFO:mfdb:Dumping table net_type
#> 2026-01-22 15:11:09.933886 INFO:mfdb:Dumping table gear
#> 2026-01-22 15:11:09.935593 INFO:mfdb:Dumping table vessel_type
#> 2026-01-22 15:11:09.936937 INFO:mfdb:Dumping table vessel_owner
#> 2026-01-22 15:11:09.938112 INFO:mfdb:Dumping table market_category
#> 2026-01-22 15:11:09.93927 INFO:mfdb:Dumping table sex
#> 2026-01-22 15:11:09.940552 INFO:mfdb:Dumping table maturity_stage
#> 2026-01-22 15:11:09.941848 INFO:mfdb:Dumping table species
#> 2026-01-22 15:11:09.946634 INFO:mfdb:Dumping table areacell
#> 2026-01-22 15:11:09.94786 INFO:mfdb:Dumping table stomach_state
#> 2026-01-22 15:11:09.949136 INFO:mfdb:Dumping table digestion_stage
#> 2026-01-22 15:11:09.950407 INFO:mfdb:Dumping table sampling_type
#> 2026-01-22 15:11:09.951573 INFO:mfdb:Dumping table population
#> 2026-01-22 15:11:09.952751 INFO:mfdb:Dumping table data_source
#> 2026-01-22 15:11:09.953927 INFO:mfdb:Dumping table index_type
#> 2026-01-22 15:11:09.95518 INFO:mfdb:Dumping table port
#> 2026-01-22 15:11:09.956339 INFO:mfdb:Dumping table trip
#> 2026-01-22 15:11:09.95751 INFO:mfdb:Dumping table tow
#> 2026-01-22 15:11:09.958793 INFO:mfdb:Dumping table vessel
#> 2026-01-22 15:11:09.959984 INFO:mfdb:Dumping table case_study
#> 2026-01-22 15:11:09.961312 INFO:mfdb:Dumping table survey_index
#> 2026-01-22 15:11:09.962519 INFO:mfdb:Dumping table division
#> 2026-01-22 15:11:09.963681 INFO:mfdb:Dumping table sample
#> 2026-01-22 15:11:09.96499 INFO:mfdb:Dumping table predator
#> 2026-01-22 15:11:09.966213 INFO:mfdb:Dumping table prey
mfdb_cs_restore(mdb_in, dump_path)
#> 2026-01-22 15:11:10.080562 INFO:mfdb:Restoring table institute
#> 2026-01-22 15:11:10.260775 INFO:mfdb:Restoring table bait_type
#> 2026-01-22 15:11:10.262416 INFO:mfdb:Restoring table net_type
#> 2026-01-22 15:11:10.263269 INFO:mfdb:Restoring table gear
#> 2026-01-22 15:11:10.432401 INFO:mfdb:Restoring table vessel_type
#> 2026-01-22 15:11:10.601172 INFO:mfdb:Restoring table vessel_owner
#> 2026-01-22 15:11:10.602517 INFO:mfdb:Restoring table market_category
#> 2026-01-22 15:11:10.603355 INFO:mfdb:Restoring table sex
#> 2026-01-22 15:11:10.770604 INFO:mfdb:Restoring table maturity_stage
#> 2026-01-22 15:11:10.93926 INFO:mfdb:Restoring table species
#> 2026-01-22 15:11:11.110844 INFO:mfdb:Restoring table areacell
#> 2026-01-22 15:11:11.112114 INFO:mfdb:Restoring table stomach_state
#> 2026-01-22 15:11:11.282916 INFO:mfdb:Restoring table digestion_stage
#> 2026-01-22 15:11:11.451135 INFO:mfdb:Restoring table sampling_type
#> 2026-01-22 15:11:11.452396 INFO:mfdb:Restoring table population
#> 2026-01-22 15:11:11.45323 INFO:mfdb:Restoring table data_source
#> 2026-01-22 15:11:11.454035 INFO:mfdb:Restoring table index_type
#> 2026-01-22 15:11:11.624203 INFO:mfdb:Restoring table port
#> 2026-01-22 15:11:11.625588 INFO:mfdb:Restoring table trip
#> 2026-01-22 15:11:11.626506 INFO:mfdb:Restoring table tow
#> 2026-01-22 15:11:11.627495 INFO:mfdb:Restoring table vessel
#> 2026-01-22 15:11:11.628331 INFO:mfdb:Restoring table case_study
#> 2026-01-22 15:11:11.79721 INFO:mfdb:Restoring table survey_index
#> 2026-01-22 15:11:11.798466 INFO:mfdb:Restoring table division
#> 2026-01-22 15:11:11.799287 INFO:mfdb:Restoring table sample
#> 2026-01-22 15:11:11.800452 INFO:mfdb:Restoring table predator
#> 2026-01-22 15:11:11.801435 INFO:mfdb:Restoring table prey

mfdb_disconnect(mdb_in)
mfdb_disconnect(mdb_out)