mfdb_helpers_mfdb_concatenate_results {mfdb}	R Documentation

MareFrame Query Utilities

Description

Aggregate data from the database in a variety of ways

Usage

mfdb_concatenate_results(...)

Arguments

...
Any number of data.frames produced by mfdb query functions with identical columns, e.g. mfdb_sample_count

Value

Given any number of data.frames from mfdb query functions with identical columns, produces a combined data.frame, similar to rbind but preserving the attributes required to produce aggregation files.

[Package mfdb version 5.2-0 Index]