mfdb_interval {mfdb}	R Documentation

MareFrame DB intervals

Description

Represent a uniform or non-uniform interval.

Usage

mfdb_interval(prefix, vect, open_ended = FALSE)
## S3 method for class 'mfdb_interval'
as.list(x, ...)

Arguments

prefix
(required) A character prefix to prepend to minimum to create list names

vect
(required) A vector representing the minimum for each group, and the maximum

open_ended
If true, the last group will include everything >= the last value in vect

x
The mfdb_interval to convert to a list for use in gadget_file

...
Ignored.

Examples

## Make groups of len40 (40--60), len60 (60--80)
g1 <- mfdb_interval("len", c(40, 60, 80))

## Use seq to make life easier
g2 <- mfdb_interval("len", seq(40, 80, by = 20))

## Convert to list representation
as.list(g2)
## Not run: 
$len40
[1] 40 60

$len60
[1] 60 80

## End(Not run)

[Package mfdb version 1.0 Index]