mfdb_sharing {mfdb}	R Documentation

MareFrame DB sharing options

Description

Alter database privileges

Usage

mfdb_share_with(mdb, user_or_role, query = TRUE, import = FALSE)

Arguments

mdb
(required) A database connection created by mfdb()

user_or_role
(required) Another database user, or a role, or 'public' to share with all users

query
Should the user be able to query the current case study?

import
Should the user be able to import more data current case study?

Details

This allows you to share case study data between users. This is most useful when using the mfdb-server. Regardless, only the owner of the schema (i.e. the user that created it) will be able to change table structure (i.e. upgrade MFDB versions).

By default nothing is shared between users.

Examples

## Not run: 
mdb <- mfdb('iceland')
mfdb_share_with(mdb, 'paul') # Allow DB user paul to query the iceland case study data
mfdb_share_with(mdb, 'mary') # Allow DB user mary to add data to the iceland case study

## End(Not run)

[Package mfdb version 5.2-0 Index]