mfdb_step_interval {mfdb}	R Documentation

MareFrame DB intervals

Description

Represent a uniform or non-uniform interval.

Usage

mfdb_step_interval(prefix, by, from = 0, to = NULL, open_ended = FALSE)
## S3 method for class 'mfdb_step_interval'
as.list(x, ...)

Arguments

prefix
(required) A character prefix to prepend to minimum to create list names

by
(required) Increment of the sequence.

from, to
Start / end of the sequence. Defaults to 0 / infinity respectively.

open_ended
If true, there should be a final group representing everything above "to".

x
An mfdb_step_interval object.

...
Nothing should be specfied (required by generic signature).

Examples

## Make groups of len0 (0--5), len5 (5--10), ... len45(45--50)
g1 <- mfdb_step_interval("len", 5, to = 50)

[Package mfdb version 1.0 Index]