mfdb_unaggregated {mfdb}	R Documentation

MareFrame DB unaggregated data

Description

Tell mfdb functions not to aggregate this column, just return all values.

Usage

mfdb_unaggregated(omitNA = FALSE)
## S3 method for class 'mfdb_unaggregated'
as.list(x, ...)

Arguments

omitNA
Skip over rows where column is NA

x
The mfdb_unaggregated to convert to a list for use in gadget_file

...
Ignored.

Examples

## Currently we can't work out what to fill a group with,
## so returns placeholders
as.list(mfdb_unaggregated())
## Not run: 
$X
[1] "X"

## End(Not run)

[Package mfdb version 1.0 Index]