ewe_model {mfdb}	R Documentation

MareFrame DB EwE/Rpath interface

Description

Transform the results of MFDB queries for use in an EwE model

Usage

ewe_stanza_group(survey_data)
ewe_stanzas(survey_data)
ewe_model(area_data, survey_data, catch_data = NULL)
ewe_diet(consumption_data)
ewe_pedigree(survey_data, catch_data = NULL)

Arguments

area_data
Results of an mfdb_area_size query, aggregating the whole area

survey_data
Results of an mfdb_sample_totalweight query, normally for one year, aggregated by the model's functional groups

catch_data
Results of an mfdb_sample_totalweight query, normally for one year, aggregated by the model's functional groups and 'vessel'

consumption_data
Results of an mfdb_stomach_preyweightratio query, aggregated by functional groups

Details

EwE requires stanzas and groups of stanzas, these are made up using the first and any other groupings in MFDB. For example, if survey_data was made with a query like mfdb_sample_totalweight(mdb, c('species', 'age'), ...), then the species will make up the generated stanza_groups, and age will make up the stanzas within those groups.

catch_data requires data that is also aggregated by vessel, this will be ignored for the purposes of deciding the stanza/stanza_group.

consumption_data treats prey groupings seperate to predator groupings, and all will be added to the diet matrix.

See mfdb_sample_totalweight for more information on how groupings can be used in queries.

Value

All return data.frame objects matching the EwE output, apart from diet which returns the diet matrix as a matrix.

Examples

# See demo/example-ewe.R for a full-length example

[Package mfdb version 5.0-0 Index]