gadget_fleetfile {mfdb}	R Documentation

Gadget fleet files

Description

Structures representing fleet file components

Usage

gadget_fleet_component(type,
        name = type,
        livesonareas = unique(data$area),
        multiplicative = 1,
        suitability = NULL,
        fleetfile = 'fleet',
        data = stop("data not provided"),
        ...)

Arguments

type
Required. Type of fleet component to create, e.g. 'totalfleet'

name
Optional. A descriptive name for the fleet component, defaults to the type.

livesonareas
Optional. Vector of area names, defaults to all unique areas in data.

multiplicative
Optional. Defaults to 1

suitability
Optional. Defaults to empty string

fleetfile
Optional. The fleet file to put the component in. Defaults to 'fleet'.

data
Required. The data.frame to use for 'amountfile'. Areas are translated into integers before adding to amountfile.

...
Extra parameters for the component, see details

Details

effortfleet requires the following extra parameters:

catchability

A list of stock names to catchability constants

quotafleet requires the following extra parameters:

quotafunction

Function name, e.g. 'simple'

biomasslevel

Vector of biomass levels

quotalevel

Vector of fishing levels

Examples

## Not run: 
    # Write a 'totalfleet' component
    gadget_dir_write(gd, gadget_fleet_component(
        'totalfleet',
        name = 'igfs',
        data = mfdb_sample_count(mdb, c('vessel'), list(
            year = 1998,
            vessel = mfdb_unaggregated(),
            step = mfdb_timestep_yearly))[[1]]))

    gadget_dir_write(gd, gadget_fleet_component(
        'effortfleet',
        name = 'barry',
        suitability = "function constant 4;",
        catchability = list(stockA=4, stockB=5),
        quotafunction = 'simple',
        biomasslevel = c(1000, 2000),
        quotalevel = c(0.1, 0.4, 0.9),
        data = mfdb_sample_count(mdb, c('vessel'), list(
            year = 1998,
            vessel = mfdb_unaggregated(),
            step = mfdb_timestep_yearly))[[1]]))

    gadget_dir_write(gd, gadget_fleet_component(
        'quotafleet',
        name = 'barry',
        suitability = "function constant 4;",
        catchability = list(stockA=4, stockB=5),
        quotafunction = 'simple',
        biomasslevel = c(1000, 2000),
        quotalevel = c(0.1, 0.4, 0.9),
        data = mfdb_sample_count(mdb, c('vessel'), list(
            year = 1998,
            vessel = mfdb_unaggregated(),
            step = mfdb_timestep_yearly))[[1]]))

## End(Not run)

[Package mfdb version 3.1-0 Index]