gadget_stockfile {mfdb}	R Documentation

Gadget stock files

Description

Structures representing a GADGET stock file

Usage

gadget_stockfile_extremes(stock_name, data)
gadget_stockfile_refweight(stock_name, data)
gadget_stockfile_initialconditions(stock_name, data)
gadget_stockfile_recruitment(stock_name, data)

Arguments

stock_name
A name, e.g. cod.imm, used as the name for the stockfile

data
A data.frame used to generate the data. See details.

Details

The columns required in the data varies depends on which function you are using.

gadget_stockfile_extremes requires age and length columns and populates minlength, minage, maxlength, maxage. The values are obtained by the grouping used, rather than the maximum values in the data. If you want the minimum and maximum from the data, query with length = NULL, age = NULL, so the table contains "all" and the grouping contains the actual minimum and maximum.

gadget_stockfile_refweight requires a length column and a mean column representing mean weight for that length group. It populates the refweightfile and dl.

gadget_stockfile_initialconditions requires area, age, length, number and mean (weight) columns. Populates initialconditions minlength, minage, maxlength, maxage, dl and the numberfile. As before, the min/max values are populated using the groupings you specify, not the min/max available data.

gadget_stockfile_recruitment requires year, step, area, age, length, number and mean (weight) columns. Populates doesrenew, minlength, maxlength, dl, numberfile.

Value

The return value is a gadget_stockfile object that can be written to the filesystem with gadget_dir_write.

Examples

# Write min/max values and refweight file 
## Not run: 
gd <- gadget_directory('./out')
imm_data <- mfdb_sample_meanweight(mdb, c('age', 'length'), list(
    age = NULL, # The age column will say 'all', but will know the min/max
    length = mfdb_step_interval('', 10, to = 100),
    species = 'COD',
    maturity_stage = '1',
    null = NULL))

# Write both min/max and refweighfile into our gadget directory
gadget_dir_write(gd, gadget_stockfile_extremes('cod.imm', imm_data[[1]]))
gadget_dir_write(gd, gadget_stockfile_refweight('cod.imm', imm_data[[1]]))

## End(Not run)

[Package mfdb version 5.0-0 Index]