mfdb {mfdb}	R Documentation

MareFrame DB class

Description

Create a class representing a connection to a MareFrame DB

Usage

    mfdb(case_study_name,
         db_params = list(),
         destroy_schema = FALSE,
         save_temp_tables = FALSE)

    mfdb_disconnect(mdb)

Arguments

case_study_name
The name of the case study you are working with, should be a valid name from mfdb::case_study.

db_params
Extra parameters to supply to DBI::dbConnect. By default it will search for a "mf" database locally, but you can override any of the parameters, in particular host, dbname, user, password. See methods?RPostgreSQL::dbConnect for more information.

destroy_schema
Optional boolean. If true, all mfdb tables will be dropped, allowing you to recreate your schema. Returns NULL, you need to call mfdb to connect and rebuild schema.

save_temp_tables
Optional boolean. If true, any temporary tables will be made permanent for later inspection.

mdb
Database connection created by mfdb().

Examples

## Not run: 
    # Connect to local DB, as the Iceland case study
    mdb <- mfdb('Iceland')

    # Connect to "mf2014" on remote DB server
    mdb <- mfdb('Iceland', db_params = list(host = "db.mareframe-fp7.org", dbname = "mf2014")

    # Stop it and tidy up
    mfdb_disconnect(mdb)

    # Destroy everything, since schema was out of date.
    mfdb('', destroy_schema = TRUE)

## End(Not run)

[Package mfdb version 4.0-0 Index]