mfdb-data {mfdb}	R Documentation

MareFrame DB Datasets

Description

Data sets representing the content of taxonomies used in the database.

Usage

case_study
gear
institute
market_category
maturity_stage
sex
species
vessel_type

Details

All of these tables represent acceptible values for use when importing data. You can see the content of an individual taxonomy at the R command line, e.g. mfdb::gear

Each of the datasets will have the following columns.

id

A numeric ID to be used internally

name

An alphanumeric ID to be used when importing and reporting data.

description

Some text describing the option.

t_group

Groups together several items to query all in one go. e.g. for institutes you can query 'NOR' to get all institutes in Norway.

The taxonomies are used in the following locations:

case_study

Possible case studies, use when connecting with mfdb()

gear, institute, vessel_type

Used to describe the dataset being imported with mfdb_import_survey()

sex, maturity_stage, species

Used for individual sample points when using mfdb_import_survey()

[Package mfdb version 4.0-0 Index]