mfdb-server {mfdb}	R Documentation

MareFrame DB server

Description

Connecting to the remote MareFrame DB server

Details

The University of Iceland host an internet-accessible Postgres database. This can be used from your local computer in much the same way as a local database, and case studies can be imported / exported locally.

Please send an an email to request access.

Connecting to the MareFrame DB server

Connection is done with the mfdb function, as with a local database. For example:-

mdb <- mfdb('iceland', db_params = list(host = 'mfdb.rhi.hi.is'))

...and available case studies will be listed if case_study_name is not provided.

mdb <- mfdb(db_params = list(host = 'mfdb.rhi.hi.is'))

When connecting to a remote database, username and password are essential. These could be specified in db_params, however, if not specified you will be prompted to enter them.

Once connected, it can be used just like any other MFDB database.

Import / Export to/from MFDB server

As well as issuing MFDB commands on the remote database, you can (depending on permissions), import and export case study examples. This is useful to:

Get example data to get up and running

Back up any data inserted locally

Share data with others

To do this, you can use a combination of pg_dump and pg_restore to copy a case study (or schema, in postgresql-speak):

pg_dump --schema=example_iceland --format-c mf | pg_restore \
    -d mf --create --clean --host=mfdb.rhi.hi.is --username=(username)

If you want to copy from the MFDB server to your computer, put the --host and --username parameters onto the pg_dump command.

Sharing case studies with other users

By default, only you are able to share case study data. However the mfdb_share_with() command allows you to let others read / modify your case study. For example:-

mfdb_share_with(mdb, 'anne')

... will allow anne to read the case study you are currently connected to.

Connecting directly to MFDB server

Finally, you can connect directly using psql if it proves useful. Note you shouldn't need to do this for general use.

psql "host=mfdb.rhi.hi.is user=(username) dbname=mf"

[Package mfdb version 5.2-0 Index]