MareFrame DB Dump / Restore

Dump / Restore entire case studies.

mfdb_cs_dump(mdb, out_location)
mfdb_cs_restore(mdb, in_location)

Arguments

mdb
(required) A database connection created by mfdb()

in_location, out_location
(required) A filesystem directory or '.tar.gz' file to dump / restore database contents.

Details

Deprecated: These commands aren't strictly necessary any more. In most situations it will be easier to use Postgres' pg_dump and pg_restore. See mfdb-package for some examples of how to do it. These functions don't offer much more functionality and much slower.

mfdb_cs_dump copies all data from the database/case-study that mdb is connected to, and writes it out to files in out_location. If this ends with 'tar.gz', then all files will be put into a tarball with the name

mfdb_cs_restore will remove any case-study data, and replaces it with the content of in_location, a directory or tarball.

Value

NULL

Examples

if (FALSE)   # NB: Requires a PostgreSQL installation, see README

# Copy case study data from one case study to another
mdb_out <- mfdb('examples')
mdb_in <- mfdb('examples-copy')

#> 2021-02-23 11:27:06 INFO:mfdb:Creating indexes

dump_path <- tempfile(fileext='.tar.gz')
mfdb_cs_dump(mdb_out, dump_path)

#> Error in mfdb_cs_dump(mdb, temp_output): object 'mdb_out' not found

mfdb_cs_restore(mdb_in, dump_path)

#> Warning: ‘/bin/tar -xf '/tmp/RtmpkeglDl/file275f4e9693c3.tar.gz' -C '/tmp/RtmpkeglDl/mfdb_cs_restore275f73024364'’ returned error code 2

#> Warning: cannot open file '/tmp/RtmpkeglDl/mfdb_cs_restore275f73024364/institute': No such file or directory

#> 2021-02-23 11:27:07 WARNING:mfdb:Rolling back transaction...

#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection