mfdb_bulk {mfdb}	R Documentation

MareFrame DB Dump / Restore

Description

Dump / Restore entire case studies.

Usage

mfdb_cs_dump(mdb, out_location)
mfdb_cs_restore(mdb, in_location)

Arguments

mdb
(required) A database connection created by mfdb()

in_location, out_location
(required) A filesystem directory or '.tar.gz' file to dump / restore database contents.

Details

mfdb_cs_dump copies all data from the database/case-study that mdb is connected to, and writes it out to files in out_location. If this ends with 'tar.gz', then all files will be put into a tarball with the name

mfdb_cs_restore will remove any case-study data, and replaces it with the content of in_location, a directory or tarball.

Examples

## Not run: 
# Copy case study data from one case study to another
mdb_out <- mfdb('Iceland')
mdb_in <- mfdb('Test')

mfdb_cs_dump(mdb_out, '/tmp/cs_dump')
mfdb_cs_restore(mdb_in, '/tmp/cs_dump')

mfdb_cs_dump(mdb_out, '/tmp/cs_dump.tar.gz')
mfdb_cs_restore(mdb_in, '/tmp/cs_dump.tar.gz')

## End(Not run)

[Package mfdb version 4.0-0 Index]