mfdb_dplyr {mfdb}	R Documentation

MareFrame DB dplyr interface

Description

Use mfdb tables with dplyr

Usage

mfdb_dplyr_survey_index(mdb, include_cols = all_cols)
mfdb_dplyr_division(mdb, include_cols = all_cols)
mfdb_dplyr_sample(mdb, include_cols = all_cols)
mfdb_dplyr_predator(mdb, include_cols = all_cols)
mfdb_dplyr_prey(mdb, include_cols = all_cols)

Arguments

mdb
An object created by mfdb()

include_cols
Any additonal columns to include in output, see details.

Details

Warning: Whilst these might be handy for exploration, there is no guarantee that code using these will continue to work from one version of MFDB to the next.

There is one function for each measurement table. By default every possible taxonomy column is included. However this is somewhat inefficient if you do not require the data, in which case specify the columns requred with include_cols. See mfdb::mfdb::taxonomy and mfdb::mfdb_cs_taxonomy for possible values.

Value

A dplyr table object, for you to do as you please.

Examples

## Not run: 
# Include as many columns as possible
mfdb_dplyr_sample(mdb)

# Only include 'data_source' and 'species' columns, as well as measurements
mfdb_dplyr_sample(mdb, c('data_source', 'species'))

## End(Not run)

[Package mfdb version 5.1-0 Index]